fix: change IOracle.sol back to recently deployed version.

+++ contracts/interfaces/IOracle.sol
@@ -2,8 +2,8 @@
 
 interface IOracle {
     /** current price for token asset. denominated in USD */
-    function getLatestAnswer(address token) external returns (int256);
+    function getLatestAnswer(address token) external returns (int);
 
     /** Readonly function providing the current price for token asset. denominated in USD */
-    function _getLatestAnswer(address token) external view returns (int256);
+    function _getLatestAnswer(address token) external view returns (int);
 }
